Revert "stringmatching renaming left overs"
This reverts commit 6cd109aeb9d2a284ea6d22ce8801969bf7c26979.
