+++ src/index.js
@@ -1,7 +1,7 @@
 import { parse } from 'acorn-loose';
 import { readFileSync, writeFileSync } from 'fs';
 import { simple, full, ancestor, fullAncestor, findNodeAt, findNodeAround, findNodeAfter } from 'acorn-walk';
-import { stringMatch } from '../src/stringMatching.js'
+import { stringMatch } from '../src/string_matching.js'
 
 const CURSOR_SYMBOL = 'Ʌ'
 /*
